Consistent package structuring for direct Terra clients
