+++ api/src/main/java/org/pmiops/workbench/iam/IamServiceImpl.java
@@ -11,6 +11,8 @@ import org.pmiops.workbench.firecloud.ApiException;
 import org.pmiops.workbench.firecloud.FirecloudApiClientFactory;
 import org.pmiops.workbench.firecloud.api.TermsOfServiceApi;
 import org.pmiops.workbench.google.CloudResourceManagerService;
+import org.pmiops.workbench.sam.SamApiClientFactory;
+import org.pmiops.workbench.sam.SamRetryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
