+++ api/src/main/java/org/pmiops/workbench/rawls/RawlsRetryHandler.java
@@ -1,28 +1,27 @@
-package org.pmiops.workbench.firecloud;
+package org.pmiops.workbench.rawls;
 
 import java.net.SocketTimeoutException;
 import java.util.logging.Logger;
 import javax.inject.Provider;
 import javax.servlet.http.HttpServletResponse;
-import org.pmiops.workbench.calhoun.ApiException;
 import org.pmiops.workbench.exceptions.ExceptionUtils;
 import org.pmiops.workbench.exceptions.WorkbenchException;
 import org.pmiops.workbench.firecloud.api.TermsOfServiceApi;
+import org.pmiops.workbench.terra.TerraServiceRetryHandler;
 import org.pmiops.workbench.utils.ResponseCodeRetryPolicy;
-import org.pmiops.workbench.utils.TerraServiceRetryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.backoff.BackOffPolicy;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CalhounRetryHandler extends TerraServiceRetryHandler<ApiException> {
+public class RawlsRetryHandler extends TerraServiceRetryHandler<ApiException> {
 
-  private static final Logger logger = Logger.getLogger(CalhounRetryHandler.class.getName());
+  private static final Logger logger = Logger.getLogger(RawlsRetryHandler.class.getName());
 
-  private static class CalhounRetryPolicy extends ResponseCodeRetryPolicy {
+  private static class RawlsRetryPolicy extends ResponseCodeRetryPolicy {
 
-    public CalhounRetryPolicy() {
-      super("Calhoun API");
+    public RawlsRetryPolicy() {
+      super("Rawls API");
     }
 
     @Override
@@ -42,7 +41,7 @@ public class CalhounRetryHandler extends TerraServiceRetryHandler<ApiException> 
         logger.log(
             getLogLevel(responseCode),
             String.format(
-                "Exception calling Calhoun API with response: %s",
+                "Exception calling Rawls API with response: %s",
                 ((ApiException) t).getResponseBody()),
             t);
       } else {
@@ -52,14 +51,14 @@ public class CalhounRetryHandler extends TerraServiceRetryHandler<ApiException> 
   }
 
   @Autowired
-  public CalhounRetryHandler(
+  public RawlsRetryHandler(
       BackOffPolicy backoffPolicy, Provider<TermsOfServiceApi> termsOfServiceApiProvider) {
-    super(backoffPolicy, new CalhounRetryPolicy(), termsOfServiceApiProvider);
+    super(backoffPolicy, new RawlsRetryPolicy(), termsOfServiceApiProvider);
   }
 
   @Override
   protected WorkbenchException convertException(ApiException exception) {
     return maybeConvertMessageForTos(exception.getCode())
-        .orElseGet(() -> ExceptionUtils.convertCalhounException(exception));
+        .orElseGet(() -> ExceptionUtils.convertRawlsException(exception));
   }
 }
