+++ api/src/main/java/org/pmiops/workbench/sam/SamRetryHandler.java
@@ -1,28 +1,28 @@
-package org.pmiops.workbench.firecloud;
+package org.pmiops.workbench.sam;
 
 import java.net.SocketTimeoutException;
 import java.util.logging.Logger;
 import javax.inject.Provider;
 import javax.servlet.http.HttpServletResponse;
+import org.broadinstitute.dsde.workbench.client.sam.ApiException;
 import org.pmiops.workbench.exceptions.ExceptionUtils;
 import org.pmiops.workbench.exceptions.WorkbenchException;
 import org.pmiops.workbench.firecloud.api.TermsOfServiceApi;
-import org.pmiops.workbench.rawls.ApiException;
+import org.pmiops.workbench.terra.TerraServiceRetryHandler;
 import org.pmiops.workbench.utils.ResponseCodeRetryPolicy;
-import org.pmiops.workbench.utils.TerraServiceRetryHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.backoff.BackOffPolicy;
 import org.springframework.stereotype.Service;
 
 @Service
-public class RawlsRetryHandler extends TerraServiceRetryHandler<ApiException> {
+public class SamRetryHandler extends TerraServiceRetryHandler<ApiException> {
 
-  private static final Logger logger = Logger.getLogger(RawlsRetryHandler.class.getName());
+  private static final Logger logger = Logger.getLogger(SamRetryHandler.class.getName());
 
-  private static class RawlsRetryPolicy extends ResponseCodeRetryPolicy {
+  private static class SamRetryPolicy extends ResponseCodeRetryPolicy {
 
-    public RawlsRetryPolicy() {
-      super("Rawls API");
+    public SamRetryPolicy() {
+      super("Sam API");
     }
 
     @Override
@@ -42,7 +42,7 @@ public class RawlsRetryHandler extends TerraServiceRetryHandler<ApiException> {
         logger.log(
             getLogLevel(responseCode),
             String.format(
-                "Exception calling Rawls API with response: %s",
+                "Exception calling Sam API with response: %s",
                 ((ApiException) t).getResponseBody()),
             t);
       } else {
@@ -52,14 +52,14 @@ public class RawlsRetryHandler extends TerraServiceRetryHandler<ApiException> {
   }
 
   @Autowired
-  public RawlsRetryHandler(
+  public SamRetryHandler(
       BackOffPolicy backoffPolicy, Provider<TermsOfServiceApi> termsOfServiceApiProvider) {
-    super(backoffPolicy, new RawlsRetryPolicy(), termsOfServiceApiProvider);
+    super(backoffPolicy, new SamRetryPolicy(), termsOfServiceApiProvider);
   }
 
   @Override
   protected WorkbenchException convertException(ApiException exception) {
     return maybeConvertMessageForTos(exception.getCode())
-        .orElseGet(() -> ExceptionUtils.convertRawlsException(exception));
+        .orElseGet(() -> ExceptionUtils.convertSamException(exception));
   }
 }
