+++ api/tools/src/main/java/org/pmiops/workbench/tools/DeleteWorkspaces.java
@@ -15,15 +15,15 @@ import org.pmiops.workbench.actionaudit.auditors.BillingProjectAuditorImpl;
 import org.pmiops.workbench.audit.ActionAuditSpringConfiguration;
 import org.pmiops.workbench.firecloud.FireCloudServiceImpl;
 import org.pmiops.workbench.firecloud.FirecloudApiClientFactory;
-import org.pmiops.workbench.firecloud.RawlsApiClientFactory;
 import org.pmiops.workbench.google.GoogleConfig;
-import org.pmiops.workbench.iam.SamApiClientFactory;
-import org.pmiops.workbench.iam.SamRetryHandler;
 import org.pmiops.workbench.impersonation.ImpersonatedFirecloudServiceImpl;
 import org.pmiops.workbench.impersonation.ImpersonatedWorkspaceService;
 import org.pmiops.workbench.impersonation.ImpersonatedWorkspaceServiceImpl;
 import org.pmiops.workbench.model.WorkspaceResponse;
+import org.pmiops.workbench.rawls.RawlsApiClientFactory;
 import org.pmiops.workbench.rawls.model.RawlsWorkspaceListResponse;
+import org.pmiops.workbench.sam.SamApiClientFactory;
+import org.pmiops.workbench.sam.SamRetryHandler;
 import org.pmiops.workbench.utils.mappers.CommonMappers;
 import org.pmiops.workbench.utils.mappers.FirecloudMapperImpl;
 import org.pmiops.workbench.utils.mappers.WorkspaceMapperImpl;
